corrected structure of the files
