+++ .github/workflows/python.yml
@@ -1,12 +1,11 @@
 name: TwitterTests
 on: [push]
 jobs:
-  Explore-GitHub-Actions:
+  tests:
     runs-on: ubuntu-latest
     steps:
       - name: Check out repository code
         uses: actions/checkout@v3
-      - run: cd Twitter-Data-Analysis
       - name: List files in the repository
         run: |
           ls ${{ github.workspace }}
