--- Twitter-Data-Analysis/tests/__init__.py
@@ -1,1 +1,0 @@
-# tests for ExtractTweet class
